==> analyses/bulkRNAseq_Analysis/06-B.4.analysis.guide.Rmd <==
<!-- Set PArameters Module -->
<!-- Set the chnkPrefix to make all chunks unique in the overall folder -->
```{r, echo=FALSE, eval=TRUE, warning=FALSE}
chnkPrefix <- "C4.guide.txt."
VersionPdfExt <- VersionPdfExt <- paste0(".",chnkPrefix,"V", gsub("-", "", Sys.Date()), ".pdf")

```


# RNA-Seq Analysis Guide

## TECHNICAL CHECKS

Quality Control (QC)

You may want to start by reviewing the quality of the underlying RNA and RNA-sequencing. This is done in the Quality control (QC) section. Things to look for are the percentage of reads that aligned to intergenic regions (indicating DNA contamination in your RNA) and the sequence duplication rate (which might be a PCR artefact introduced during sample amplification and may affect the complexity of the sequenced library).



```{r, eval=T, results="asis"}
cat(paste0("A summary of this experiment and its results so far is summarized in an editable [report](https://biologic.crick.ac.uk/",Obio@parameterList$project_id,"/report.php). The purpose of this section is to provide a brief summary of the experiment as well as of its results to a person who is interested in your dataset in the distant future. Ideally, this section contains all information necessary to understand the experiment as well as its results so far. You may download this presentation via the link at the bottom of the page, edit it in whatever way you deem suitable and send it back to bioinformatics for updating. You may also want to add a powerpoint presentation outlining the rationale for this experiment as a [slide show](https://biologic.crick.ac.uk/",Obio@parameterList$project_id,"/about.php)."))
```
