==> analyses/bulkRNAseq_Analysis/08-B.6.DEseq2.PCA.Rmd <==
<!-- Set Parameters Module -->
<!-- Set the chnkPrefix to make all chunks unique in the overall folder -->
```{r, echo=FALSE, eval=TRUE, warning=FALSE}
chnkPrefix <- "edit.RSEM.tpm."
VersionPdfExt <- VersionPdfExt <- paste0(".",chnkPrefix,"V", gsub("-", "", Sys.Date()), ".pdf")

```

# DESeq2 Analysis
## Create DESeq2 object
```{r, echo=TRUE, eval=TRUE, warning=FALSE, results="asis"}
###############################################################################
## Create and attach DDS object                                              ##

if (is.null( Obio@parameterList$parallelProcessing) || length(Obio@parameterList$parallelProcessing) == 0){
     Obio@parameterList$parallelProcessing <- FALSE
}

Obio <- biologicSeqTools2::createDdsObject(Obio)

## Done creating and attaching DDS objecs                                    ##
###############################################################################
```

## Variation and PCA
```{r, echo=TRUE, eval=TRUE, warning=FALSE, results="hide", message=FALSE}
###############################################################################
## Determine variation in the data set                                       ##

## This is a tempory fix until the PCA function has been fully equipped for
## batch correction
saveMode <- Obio@parameterList$batchMode
Obio@parameterList$batchMode <- FALSE

Obio <- biologicSeqTools2::addCoVar(
    obj = Obio,
    avgCountCutOffperSample = 1,
    selectionColName = "aboveCutOff",
    dfBaseData = Obio@DESeqNormReadCountsTable,
    rowNameID = Obio@parameterList$primaryAlignmentGeneID
    #options: "DEseq2RV" or "CoVar"
)

## Results are in slot Obio@dataTableList$dfRowVar ##

## Diagnostic plot ##
# library(ggplot2)
# countCutOff <- 0
# pCoVar <- ggplot(
#     data=Obio@dataTableList$dfRowVar[Obio@dataTableList$dfRowVar$avgCountsPerGenePerSample > countCutOff ,],
#     aes(
#         x=CoVar,
#         y=DEseq2RV
#     )) + geom_point(
#     )  + labs(title = "Variation" ,x = "CoVar", y = "DeSeq2Var"
#     ) +  theme(
#         axis.text.y   = element_text(size=8),
#         axis.text.x   = element_text(size=8),
#         axis.title.y  = element_text(size=8),
#         axis.title.x  = element_text(size=8),
#         axis.line = element_line(colour = "black"),
#         panel.border = element_rect(colour = "black", fill=NA, size=1),
#         plot.title = element_text(hjust = 0.5, size = 12)
#     )

## Done determine variation in the data set                                  ##
###############################################################################

###############################################################################
## Perform PCA, MV-analysis, and Clusterdendrogram                           ##

## Select elements for PCA explicityly ##
dfSel <- Obio@dataTableList$dfRowVar

## Remove low count values ##
dfSel <- dfSel[dfSel$avgCountsPerGenePerSample >= 1, ]

## Order by variation
dfSel <- dfSel[order(dfSel$DEseq2RV, decreasing = TRUE),]
# dfSel <- dfSel[order(dfSel$CoVar, decreasing = TRUE),]

rowSelVec <- as.vector(
    dfSel[1:Obio@parameterList$NtopGenes,Obio@parameterList$primaryAlignmentGeneID]
)

Obio@dataTableList[["Ntop4pcaGeneSelection"]] <- rowSelVec



Obio <- doPCA(
    obj = Obio,
    pcaDimensionsToInvestigate = c(1:7),
    Ntop4pca = Obio@parameterList$NtopGenes, #500,
    avgCountCutOffperSample = 0,
    pcaSelectionVec = rowSelVec
)

Obio@parameterList$batchMode <- saveMode
## Do linear fittings to available dimensions ##


# depdendent Variation mode: Variables can be dependent (individual fits)


if (length(unique(Obio@dfDesign$dataseries)) > 1){
    independentDesignColSector <- c(
        "dataseries"
    )
} else {
    independentDesignColSector <- as.vector(NULL, mode = "character")
}

if (length(Obio@dfDesign$sample.group) > 1){
    independentDesignColSector <- c(
        independentDesignColSector,
        "sample.group"
    )
}

if (length(Obio@dfDesign$replicate) > 1){
    independentDesignColSector <- c(
        independentDesignColSector,
        "replicate"
    )
}

pos <- grep("^f_", names(Obio@dfDesign))
if (length(pos) > 0){
  independentDesignColSector <- c(
    names(Obio@dfDesign)[grep("^f_", names(Obio@dfDesign))]
  )

} else if (length(unique(Obio@dfDesign$dataseries)) > 1){
    independentDesignColSector <- c(
      "dataseries"
    )
}


# indedendentVariation mode: Variables are required to be independent
## Do linear variable fittings to PCA ##
# variables need to be independent > culmulative fit #
Obio <- biologicSeqTools2::doLinearFittings(
    obj = Obio,
    designColSelector = independentDesignColSector,
    mode = "independentVariation", ## "independentVariation" or "dependentVariation"
    Ntop4pca = Obio@parameterList$NtopGenes,
    plotname = "P2_independentVariables"
)



## Do linear variable fittings to PCA ##

# individual fit
LRTvec <- names(Obio@dfDesign)[grep("f_", names(Obio@dfDesign))]
#LRTvec <- LRTvec[LRTvec != "LRT_replicate"]
#LRTvec <- "LRT_timepoint"

dependentDesignColSelector<- c(
    independentDesignColSector,
    LRTvec,
    names(Obio@dfDesign)[grep("comp_", names(Obio@dfDesign))],
    names(Obio@dfDesign)[grep("LRT_", names(Obio@dfDesign))]
)
Obio <- biologicSeqTools2::doLinearFittings(
    obj = Obio,
    designColSelector = dependentDesignColSelector,
    mode = "dependentVariation", ## "independentVariation" or "dependentVariation"
    Ntop4pca = Obio@parameterList$NtopGenes,
    plotname = "P2_dependentVariables"
)


## Create Plot dendrogram ##
# Obio <- createSampleDendrogram(
#     obj = Obio,
#     Ntop4pca = Obio@parameterList$NtopGenes,
#     plotname = "dendrogram10000"
# )

###############################################################################
## Do differential gene expression analysis                                  ##
#Obio@parameterList$batchMode <- F
## LRT tests for multiple sample groups ##

# if (!is.null(Obio@projectDetailList$DEseq2_DGE_result_folder) && Obio@projectDetailList$DEseq2_DGE_result_folder != ""){
#     mode = "load_DGE_from_file"
#     Obio <- loadDESeq2outputFromFile(
#         Obio,
#         replace = TRUE
#     )
# } else {


    
#}


############################################################
## Function load DGE results from file
loadDESeq2outputFromFile <- function(
    Obio,
    DEseq2resultDir = "Obio@parameterList$DEseq2External_DGE",
    replace = FALSE,
    mode = NULL # Can be DGE or LRT
){
    
    if (is.null(mode)){
        if (DEseq2resultDir == "DEseq2External_LRT"){
            mode = "LRT"
        } else {
            mode = "DGE"
        }
    }
  
    allfiles <- paste0(DEseq2resultDir, "/", list.files(DEseq2resultDir))
    #allfiles <- allfiles[grep(".txt", allfiles)]
    allfiles <- gsub("//", "/", allfiles)
    allfiles <- sort(allfiles)
    contrastNames <- gsub(gsub("//", "/", paste0(DEseq2resultDir, "/")), "", allfiles)
    contrastNames <- gsub(".txt", "", contrastNames)





    ###############################################################################
    ## For this project only: Add ENSMUSG column                                 ##
    dfAnno <- Obio@dfGeneAnnotation
    dfAnno <- unique(
      Obio@dfGeneAnnotation[,c(Obio@parameterList$primaryAlignmentGeneID, Obio@parameterList$geneIDcolumn)]
    )
    
    ## Done adding ENSMUSG column                                                ##
    ###############################################################################
    for (i in 1:length(allfiles)){
      colName <- contrastNames[i]
      res <- read.delim(allfiles[i], header = T, sep="\t")
      
      
      ## In case gene_ids were saved in row names
      pos <- grep("gene_id", names(res))
      if (length(pos) == 0){
          res[["gene_id"]] <- row.names(res)  
      } else if (length(pos) == 1){
          row.names(res) <- res$gene_id
      }
      
      ################################
      ## This project only 
      
      
      selVec <- c("gene_id", "baseMean", "log2FoldChange", "lfcSE",  "pvalue", "padj")
      
      if (!sum(selVec %in% names(res)) == length(selVec)){
          stop(paste0("Check DESeq2 input files. Mandantory columns: ", paste0(selVec, collapse = ", ")))
      }
      
      res <- unique(res[,selVec])
      
      res <- res[!(duplicated(res[,"gene_id"])),]
      
      ## Remove all log-fcs with an NA padj
      res <- res[!is.na(res$padj),]
      res <- res[!is.na(res$log2FoldChange),]
      
      
      res[is.na(res)] <- 0
      res <- res[res$baseMean > 0,]
      
      # Plus one to avoid negative log2 baseMeans
      res$baseMean <- log2((res$baseMean + 1))
      ## Done 
      ################################
      
      
      #row.names(res) <- res[,Obio@parameterList$primaryAlignmentGeneID]
      
      names(res) = paste(names(res), colName, sep="_")
      names(res) <- gsub(paste0("gene_id_", colName), "gene_id", names(res))
      names(res) <- gsub("^gene_id$", Obio@parameterList$primaryAlignmentGeneID, names(res))
      #res[[Obio@parameterList$primaryAlignmentGeneID]] = rownames(res)
      
      ## log2 the base mean for lrt applications
      
      
      if (mode == "DGE"){
          names(res) = gsub("log2FoldChange", "logFC", names(res))
          names(res) = gsub(
            "logFC",
            paste("contrast_D", i, "_logFC", sep=""),
            names(res)
          )
          
          names(res) = gsub(
            "padj",
            paste("contrast_D", i, "_padj", sep=""),
            names(res)
          )
          
          names(res) = gsub(
            "stat",
            paste("contrast_D", i, "_stat", sep=""),
            names(res)
          )
          
          names(res) = gsub(
            "baseMean",
            paste("contrast_D", i, "_lg2BaseMean", sep=""),
            names(res)
          )
          
          #Remove all rows without a padj
          padj.col = grep("padj", names(res))[1]
          res[,padj.col][is.na(res[,padj.col])] = ""
          res = res[res[,padj.col] != "", ]
          res[,padj.col] <- as.numeric(res[,padj.col])
          
          ## Add log10p column ##
          padj  <- names(res)[grep("_padj_", names(res))]
          lg10p <- gsub("padj", "lg10p", padj)
          
          for (z in 1:length(padj)){
            preprocess <- as.numeric(res[,padj[z]])
            minNum <- min(preprocess[preprocess != 0])
            preprocess[preprocess == 0] <- minNum
            
            # if (length(grep("padj_LRT", padj[i])) > 0){
            #     preprocess <- as.numeric(res[,padj[z]])
            #     minNum <- min(preprocess[preprocess != 0])
            #     preprocess[preprocess == 0] <- minNum
            # } else {
            #     preprocess <- as.numeric(res[,padj[z]])
            # }
            
            temp <- -1*log10(preprocess)
            #temp[temp >= 50] = 50
            res[,lg10p[z]] <- temp
          }
          
          col.vector = c(
            Obio@parameterList$primaryAlignmentGeneID,
            names(res)[grep("contrast", names(res))]
          )
          
          res = res[,col.vector]
          
          ## Make all numeric columns numeric ##
          res[,grep("contrast_", names(res))] <- apply(res[,grep("contrast_", names(res))], 2, as.numeric)
      }
      
      if (mode == "LRT"){
               #res[[Obio@parameterList$primaryAlignmentGeneID]] = rownames(res)

                res$stat <- NULL

                
                names(res) = gsub(
                    "baseMean",
                    paste0("contrast_L_lg2BaseMean_"),
                    names(res)
                )

                names(res) = gsub(
                    "padj",
                    paste0("contrast_L_padj_"),
                    names(res)
                )



                #Remove all rows without a padj
                padj.col = grep("padj", names(res))[1]
                res[,padj.col][is.na(res[,padj.col])] = ""
                res = res[res[,padj.col] != "", ]
                res[,padj.col] <- as.numeric(res[,padj.col])

                ## Add log10p column ##
                padj  <- names(res)[grep("_padj_", names(res))]
                lg10p <- gsub("padj", "lg10p", padj)

                for (z in 1:length(padj)){
                    preprocess <- as.numeric(res[,padj[z]])
                    minNum <- min(preprocess[preprocess != 0])
                    preprocess[preprocess == 0] <- minNum

                    # if (length(grep("padj_LRT", padj[i])) > 0){
                    #     preprocess <- as.numeric(res[,padj[z]])
                    #     minNum <- min(preprocess[preprocess != 0])
                    #     preprocess[preprocess == 0] <- minNum
                    # } else {
                    #     preprocess <- as.numeric(res[,padj[z]])
                    # }

                    temp <- -1*log10(preprocess)
                    #temp[temp >= 50] = 50
                    res[,lg10p[z]] <- temp
                }

                col.vector = c(
                    Obio@parameterList$primaryAlignmentGeneID,
                    names(res)[grep("contrast", names(res))]
                )

                res = res[,col.vector]

                ## Make all numeric columns numierc ##
                ## Make all numeric columns numierc ##
                res[,grep("contrast_", names(res))] <- apply(res[,grep("contrast_", names(res))], 2, as.numeric)
      } # end LRT mode
      
      if (i == 1){
        dfContrastTable <- res
      } else {
        dfContrastTable <- merge(
          dfContrastTable,
          res,
          by.x = Obio@parameterList$primaryAlignmentGeneID,
          by.y = Obio@parameterList$primaryAlignmentGeneID,
          all = TRUE
        )
        dfContrastTable[is.na(dfContrastTable)] <- 0
      }
    }
    
    
    # head(dfContrastTable)
    
    dfContrastTable<- dfContrastTable[rowSums(dfContrastTable[,2:ncol(dfContrastTable)]) != 0, ]
    names(dfContrastTable) <- gsub("__", "_", names(dfContrastTable))
    
    if (mode == "LRT"){
        if (replace){
            Obio@DEseq2LRTtable <- data.frame(NULL)
            Obio@DEseq2LRTtable <- dfContrastTable
        } else if (nrow(Obio@DEseq2LRTtable) > 0){
            Obio@DEseq2LRTtable <- merge(
                Obio@DEseq2LRTtable,
                dfContrastTable,
                by.x = Obio@parameterList$primaryAlignmentGeneID,
                by.y = Obio@parameterList$primaryAlignmentGeneID,
                all = TRUE
            )
            Obio@DEseq2LRTtable[is.na(Obio@DEseq2LRTtable)] <- 0 
        } else {
            Obio@DEseq2LRTtable <- dfContrastTable
        }
    }
    
    if (mode == "DGE"){
        if (replace){
            Obio@DEseq2contrastTable <- data.frame(NULL)
            Obio@DEseq2contrastTable <- dfContrastTable
        } else if (nrow(Obio@DEseq2contrastTable) > 0){
            Obio@DEseq2contrastTable <- merge(
                Obio@DEseq2contrastTable,
                dfContrastTable,
                by.x = Obio@parameterList$primaryAlignmentGeneID,
                by.y = Obio@parameterList$primaryAlignmentGeneID,
                all = TRUE
            )
            Obio@DEseq2contrastTable[is.na(Obio@DEseq2contrastTable)] <- 0 
        } else {
            Obio@DEseq2contrastTable <- dfContrastTable
        }
    }
    
    
    
    
    if (mode == "LRT"){
      print("DESeq2 results loaded into Obio@DEseq2LRTtable")
    } else {
      print("DESeq2 results loaded into Obio@DEseq2contrastTable")
    }
    
    return(Obio)
}

## Done with function                                                        ##
###############################################################################


    
###############################################################################
## Add or calculate LRT results                                              ##
if (!is.null(Obio@parameterList$calculate_DGE) && is.logical(Obio@parameterList$calculate_DGE)) {
    calculate_LRT <- Obio@parameterList$calculate_LRT
} else {
    calculate_LRT <- TRUE 
}




pos <- grep("LRT_", names(Obio@dfDesign))
if (length(pos) == 0){
    calculate_LRT <- FALSE
}

if (calculate_LRT){
  dfDGE <- Obio@dfModel
    dfDGE <- dfDGE[dfDGE$test == "LRT",]
  
    if (nrow(dfDGE) > 0){
        Obio <- biologicSeqTools2::LRTanalysis(
            obj = Obio,
            createNewResultTable = TRUE
        )
    }
} 

## Add additional, external LRT files, if available
lrtFilesToAdd <- NULL
if (!is.null(Obio@parameterList$DEseq2External_LRT)){
    lrtFilesToAdd <- list.files(Obio@parameterList$DEseq2External_LRT)
}    
    
if (length(lrtFilesToAdd) > 0){
    Obio <- loadDESeq2outputFromFile(
        Obio,
        replace = TRUE,
        mode = "LRT",
        DEseq2resultDir = Obio@parameterList$DEseq2External_LRT
    )
} 


## Done adding and/or calculating LRT results                                 ##
################################################################################
  

################################################################################
## Add calculate DGE results                                                  ##

if (!is.null(Obio@parameterList$calculate_DGE) && is.logical(Obio@parameterList$calculate_DGE)) {
    calculate_DGE <- Obio@parameterList$calculate_DGE
} else {
    calculate_DGE <- TRUE  
}


pos <- grep("comp_", names(Obio@dfDesign))
if (length(pos) == 0){
    calculate_DGE <- FALSE
}

if (calculate_DGE){
  ## Pairwise differential gene expression ##
    ## If working on projects prior to fall 2018, make sure
    ## Obio@parameterList$DEseq2betaPrior is set to true.
    dfDGE <- Obio@dfModel
    dfDGE <- dfDGE[dfDGE$test == "Wald",]

    if (nrow(dfDGE) > 0){
        Obio <- biologicSeqTools2::DGEanalysis(
            obj = Obio,
            createNewResultTable = TRUE
        )
    }
  
}    

## If mode is not calculate DGE, a dge table needs to be provided
dgeFilesToAdd <- NULL
if (!is.null(Obio@parameterList$DEseq2External_DGE)){
    dgeFilesToAdd <- list.files(Obio@parameterList$DEseq2External_DGE)
}    
    
if (length(dgeFilesToAdd) > 0){
    Obio <- loadDESeq2outputFromFile(
        Obio,
        replace = FALSE,
        mode = "DGE",
        DEseq2resultDir = Obio@parameterList$DEseq2External_DGE
    )
}     

## Done adding and/or calculating LRT results                                 ##
################################################################################



## Done differential gene expresison analysis                                ##
###############################################################################


```
